Remove login component
